requirements.txt
opencv-python>=4.4.0.42
numpy>=1.19.1
pillow>=7.2.0
tesserocr~=2.5.1